recommendation_api/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
mysql-connector-python
